+++ test/src/log/logger_provider_test.dart
@@ -11,24 +11,8 @@ import 'package:batch/src/log/logger.dart';
 import 'package:batch/src/log/logger_provider.dart';
 
 void main() {
-  test('Test LoggerProvider before load logger', () {
-    expect(() => trace, throwsA(isA<NoSuchMethodError>()));
-    expect(() => debug, throwsA(isA<NoSuchMethodError>()));
-    expect(() => info, throwsA(isA<NoSuchMethodError>()));
-    expect(() => warn, throwsA(isA<NoSuchMethodError>()));
-    expect(() => error, throwsA(isA<NoSuchMethodError>()));
-    expect(() => fatal, throwsA(isA<NoSuchMethodError>()));
-  });
-
-  test('Test LoggerProvider after load logger', () {
+  test('Test LoggerProvider', () {
     final logger = Logger.loadFrom(config: LogConfiguration());
-
-    expect(trace, isNotNull);
-    expect(debug, isNotNull);
-    expect(info, isNotNull);
-    expect(warn, isNotNull);
-    expect(error, isNotNull);
-    expect(fatal, isNotNull);
 
     expect(log.trace, isNotNull);
     expect(log.debug, isNotNull);
@@ -38,11 +22,5 @@ void main() {
     expect(log.fatal, isNotNull);
 
     expect(log == logger, isTrue);
-    expect(trace == logger.trace, isTrue);
-    expect(debug == logger.debug, isTrue);
-    expect(info == logger.info, isTrue);
-    expect(warn == logger.warn, isTrue);
-    expect(error == logger.error, isTrue);
-    expect(fatal == logger.fatal, isTrue);
   });
 }